started to reorganize the code to find empty colums in rectify
